% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog_slow.R
\name{mylog_slow}
\alias{mylog_slow}
\title{Title}
\usage{
mylog_slow(x)
}
\arguments{
\item{x}{a vector of length n}
}
\value{
a vector of length n
}
\description{
Title
}
\examples{
 mylog_slow(c(10, exp(2), 2))
}
